+++ src/acs/middleware.ts
@@ -3,8 +3,8 @@ import { initAuthZ, authZ } from './authz';
 /**
  * Koa middleware using the BMSLSA implementation for `iam-authn`.
  */
-export const acsClientMiddleware = () => {
-  initAuthZ();
+export const acsClientMiddleware = (config?: any) => {
+  initAuthZ(config);
 
   return async (ctx: any, next) => {
     ctx.authZ = authZ;