+++ src/index.ts
@@ -1,2 +1,5 @@
 export * from './acs/resolver';
+export * from './acs/authz';
+export * from './config';
 export * from './acs/middleware';
+export * from './acs/interfaces';