+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 0.0.2 (February 5th, 2020)
+
+- modified initAuthZ for providing config
+- added execute action
+
 ### 0.0.1 (January 29th, 2020)
 
 Initial share.